Refactor: move sync BloC/Cubit tests to mimic /lib
